Fix use-after-consume logic for more complex flows, and standardized numeric literals

+++ std/random.kotlin.sv
@@ -1,9 +1,9 @@
 // Kotlin defines for random [backend-define-handler].
 
 define handler DefaultRandom of Random {
-    define fn random() -> Float {
+    define fn random() -> Double {
         inline: ``
-        kotlin.random.Random.nextFloat()
+        kotlin.random.Random.nextDouble()
         ``
     }
 }
